move manager ok

+++ Core/Inc/dino/move_manager.h
@@ -6,7 +6,7 @@
 
 int move_manager_init(void);
 
-int move_manager_move_element(element_id_t element_id, int x, int y, float speed);
+int move_manager_move_element(element_id_t element_id, int x, int y, int speed);
 
 int move_manager_loop(void);
